requirements.txt
defusedxml==0.6.0
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.5.0
numpy==1.18.2
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
xlrd==1.2.0
xlwt==1.3.0
